+++ src/api/models/cat.go
@@ -1,0 +1,10 @@
+package models
+
+import "github.com/jinzhu/gorm"
+
+
+type Cat struct {
+	gorm.Model
+	Name string `json:"name"`
+	Type string `json:"type"`
+}
