+++ src/api/models/model.go
@@ -1,0 +1,2 @@
+package models
+
