+++ src/db/db.go
@@ -29,9 +29,12 @@ func InitDB() {
 	var err error
 	config := getConfig()
 	db, err = gorm.Open(config.Db())
+	defer db.Close()
+
 	if err != nil {
 		panic(err.Error())
 	}
+	//TODO:存在しているものはここでmigrateする
 	db.AutoMigrate(&models.Dog{})
 }
 
@@ -50,4 +53,15 @@ func (d dbConfig) DSN() string {
 
 func (c config) Db() (string, string) {
 	return c.Database.Driver, c.Database.DSN()
+}
+
+func OpenDB() *gorm.DB {
+	var err error
+	config := getConfig()
+	db, err = gorm.Open(config.Db())
+	if err != nil {
+        panic("failed to connect database.")
+    }
+	db.LogMode(true)
+	return db
 }
